+++ src/EnrolmentWorkflow/Rule/Language/LanguageFormGUI.php
@@ -2,7 +2,7 @@
 
 namespace srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\Language;
 
-use srag\CustomInputGUIs\SrUserEnrolment\MultiSelectSearchInputGUI\MultiSelectSearchInputGUI;
+use srag\CustomInputGUIs\SrUserEnrolment\MultiSelectSearchNewInputGUI\MultiSelectSearchNewInputGUI;
 use srag\CustomInputGUIs\SrUserEnrolment\TabsInputGUI\MultilangualTabsInputGUI;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\AbstractRuleFormGUI;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\RuleGUI;
@@ -41,7 +41,7 @@ class LanguageFormGUI extends AbstractRuleFormGUI
 
         $this->fields = array_merge($this->fields, [
             "languages" => [
-                self::PROPERTY_CLASS    => MultiSelectSearchInputGUI::class,
+                self::PROPERTY_CLASS    => MultiSelectSearchNewInputGUI::class,
                 self::PROPERTY_OPTIONS  => MultilangualTabsInputGUI::getLanguages(),
                 self::PROPERTY_REQUIRED => true
             ]